requirements.txt
aioredis==1.2.0
amqp==2.5.1
asgiref==3.2.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.7
django-crispy-forms==1.7.2
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
kombu==4.6.4
more-itertools==7.2.0
msgpack==0.6.2
Pillow==6.1.0
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.7
vine==1.3.0
xlrd==1.2.0
XlsxWriter==1.1.8
zipp==0.6.0
zope.interface==4.6.0
factory_boy==3.2.1
Faker==8.1.1
